apps/qod/requirements.txt
requests==2.28.2
SQLAlchemy==2.0.4
Pillow==9.4.0
opencv-python==4.7.0.68